+++ src/main/java/insane96mcp/iguanatweaksreborn/modifier/Modifiers.java
@@ -2,6 +2,7 @@ package insane96mcp.iguanatweaksreborn.modifier;
 
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.fml.ModList;
 
 import java.lang.reflect.Type;
 import java.util.HashMap;
@@ -17,6 +18,8 @@ public class Modifiers {
         registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "correct_biome"), CorrectBiomeModifier.class);
         registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "wrong_biome"), WrongBiomeModifier.class);
         registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "age"), AgeModifier.class);
+        if (ModList.get().isLoaded("sereneseasons"))
+            registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "season"), SeasonModifier.class);
     }
 
     public static Type registerModifier(ResourceLocation id, Type modifier) {