+++ changelog.md
@@ -4,6 +4,7 @@
 * Fog
   * Added fog like pre-1.7.2 which starts closer to player. Also fog starts even closer if rains
 * Added gamerule to disable experience (also disables experience bar)
+* Ported season plant growth modifier from ITE
 
 ## 3.12.2
 * Startup crash fix